+++ lib/teamlab/modules/project.rb
@@ -31,22 +31,5 @@ module Teamlab
     def initialize
       @request = Teamlab::Request.new('project')
     end
-
-    # region Import
-
-    def get_import_status
-      @request.get(['import'])
-    end
-
-    def add_importing_url_to_queue(company_url, username, password, import_closed, disable_notifications, import_users_as_collaborators, projects = [])
-      @request.post(%w[import], url: company_url, userName: username, password: password, importClosed: import_closed,
-                                disableNotifications: disable_notifications, importUsersAsCollaborators: import_users_as_collaborators, projects: projects)
-    end
-
-    def get_projects_for_import(url, username, password)
-      @request.post(%w[import projects], url: url, userName: username, password: password)
-    end
-
-    # endregion
   end
 end