+++ Changelog.md
@@ -58,6 +58,7 @@
 * Remove deprecated `Crm#delete_contact_address`
 * Remove deprecated `Crm#save_crm_settings`
 * Rename `Crm#get_settings` to `Crm#get_invoice_settings`
+* Remove Basecamp import methods
 
 ## 0.6
 ### New features